requirements.txt
appdirs==1.4.4
asgiref==3.2.10
backcall==0.2.0
boto==2.49.0
boto3==1.14.33
botocore==1.17.33
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
class-registry==2.1.2
colorama==0.4.3
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-filter==2.3.0
django-smtp-ssl==1.0
django-storages==1.9.1
djangorestframework==3.12.2
docutils==0.15.2
filelock==3.0.12
filters==1.3.2
gunicorn==20.0.4
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pickleshare==0.7.5
postgres==3.0.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
s3chunkuploader==0.14
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.2
urllib3==1.25.10
virtualenv==20.0.29
wcwidth==0.2.5
wfastcgi==3.0.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipfile36==0.1.3
